requirements.txt
Flask~=2.2.2
pandas~=1.4.3
boto3~=1.24.61
python-dotenv~=0.20.0
mysql-connector-python~=8.0.30
redis~=4.3.4